+++ QBert/Ugg_Wrongway.cpp
@@ -198,7 +198,7 @@ void Ugg_Wrongway::CheckCollision()
 	rect.x = transform->GetPosition().x;
 	rect.y = transform->GetPosition().y;
 	//make it a square, so the player only has to worry about the landing
-	rect.h = rect.w = m_CharacterWidth;
+	rect.h = rect.w = (float)m_CharacterWidth;
 	if (Utility::IsOverlappingRectangle(rect, playerRect))
 	{
 		m_cmdPlayerDeath->Execute();
